CEXT-6123: Create App Builder Actions to Test Out-of-Process Total Collector for Sales Rules
