+++ app.config.yaml
@@ -90,7 +90,7 @@ application:
               raw-http: true
               final: true
           tiered-quantity-discount:
-            function: actions/tiered-quantity-discount/index.js
+            function: actions/total-collector-discounts/tiered-quantity-discount/index.js
             web: 'yes'
             runtime: nodejs:22
             inputs:
@@ -102,7 +102,7 @@ application:
               raw-http: true
               final: true
           tiered-category-discount:
-            function: actions/tiered-category-discount/index.js
+            function: actions/total-collector-discounts/tiered-category-discount/index.js
             web: 'yes'
             runtime: nodejs:22
             inputs:
@@ -114,7 +114,7 @@ application:
               raw-http: true
               final: true
           category-based-discount:
-            function: actions/category-based-discount/index.js
+            function: actions/total-collector-discounts/category-based-discount/index.js
             web: 'yes'
             runtime: nodejs:22
             inputs:
@@ -126,7 +126,7 @@ application:
               raw-http: true
               final: true
           cheapest-item-discount:
-            function: actions/cheapest-item-discount/index.js
+            function: actions/total-collector-discounts/cheapest-item-discount/index.js
             web: 'yes'
             runtime: nodejs:22
             inputs:
@@ -138,7 +138,7 @@ application:
               raw-http: true
               final: true
           expensive-item-discount:
-            function: actions/expensive-item-discount/index.js
+            function: actions/total-collector-discounts/expensive-item-discount/index.js
             web: 'yes'
             runtime: nodejs:22
             inputs:
@@ -150,7 +150,7 @@ application:
               raw-http: true
               final: true
           cheapest-quantity-discount:
-            function: actions/cheapest-quantity-discount/index.js
+            function: actions/total-collector-discounts/cheapest-quantity-discount/index.js
             web: 'yes'
             runtime: nodejs:22
             inputs:
@@ -162,7 +162,7 @@ application:
               raw-http: true
               final: true
           step-price-discount:
-            function: actions/step-price-discount/index.js
+            function: actions/total-collector-discounts/step-price-discount/index.js
             web: 'yes'
             runtime: nodejs:22
             inputs:
@@ -174,7 +174,7 @@ application:
               raw-http: true
               final: true
           multi-condition-discount:
-            function: actions/multi-condition-discount/index.js
+            function: actions/total-collector-discounts/multi-condition-discount/index.js
             web: 'yes'
             runtime: nodejs:22
             inputs:
@@ -185,8 +185,8 @@ application:
               require-adobe-auth: false
               raw-http: true
               final: true
-          tiered_total_spend_discount:
-            function: actions/tiered_total_spend_discount/index.js
+          tiered-total-spend-discount:
+            function: actions/total-collector-discounts/tiered-total-spend-discount/index.js
             web: 'yes'
             runtime: nodejs:22
             inputs:
